ユーザー名とアイコンを表示するコンポーネントを作成
- コンポーネント<UserProfile>から<UserAvatar>に名前変更
